refactor: 💡 move generate component and generators service to lib (#53)
